feat: add music management and conversion system
Put lossless or lossy music in ~/Music/PURE
Run regen.py to convert lossless audio to MP3 V0 or copy lossy audio as is.
The newly converted or copied files will be located in ~/Music/LITE.
This LITE folder is ideal for easy sync to portable players/phone/etc.

regen.py only converts or copies if there is no lossy equivalent of a file
in PURE, or if there are changes in a PURE file such as cover art/tags.
This makes it ideal for a cronjob.

New .alias added: regen. It simply calls ~/.local/bin/regen.py.